move LessonTitleCard files to new folder
