+++ components/LessonTitleCard/lessonTitleCard.module.scss
